+++ src/main/java/net/silentchaos512/gems/block/supercharger/SuperchargerPillarStructure.java
@@ -1,6 +1,5 @@
 package net.silentchaos512.gems.block.supercharger;
 
-import lombok.Getter;
 import net.minecraft.item.Item;
 import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.tags.Tag;
@@ -11,12 +10,16 @@ import java.util.*;
  * Currently just used by JEI? Could possibly use this to simplify tile entity code somewhat.
  */
 public class SuperchargerPillarStructure {
-    @Getter private final int tier;
+    private final int tier;
     private final List<Ingredient> list = new ArrayList<>();
 
     public SuperchargerPillarStructure(int tier, Collection<Tag<Item>> blocks) {
         this.tier = tier;
         blocks.stream().map(Ingredient::fromTag).forEach(this.list::add);
+    }
+
+    public int getTier() {
+        return tier;
     }
 
     public List<Ingredient> getBlocks() {