+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- (API) Replace Lombok getters with normal getters, removed some less useful getters
+
 ## [3.5.1] - 2020-02-19
 ### Added
 - Configs to control number of gem types per biome in the overworld [#448]